Threading code - complete but untested
Have not tested the threading code. Also won't compile due to argmuent
type mismatch. Found out pthreads is a unix based thing, so we will have
to use a Win32 port of it. Updated the VS2015 project to reflect this.
Will copy the library over to other desktop later. I have the files in a
zip folder if needed.