Threading implemented, still untested
All the code for threading should be in place. TODO: Still need to
modify the call back to use these functions, and test with glasses. I
will include link to the external win32-pthreads library in a future
commit (add headers, .lib and .dll for windows). Version I got was for
VS 2013...really hope it works.